Moved input event management to a separate module.
