+++ linux/Makefile
@@ -32,7 +32,7 @@ all: $(TARGETS)
 clean:
 	rm -f *.o $(TARGETS)
 
-zeST: setup.o floppy.o floppy_img.o ikbd.o sil9022a.o osd.o zui.o menu.o
+zeST: setup.o floppy.o floppy_img.o ikbd.o sil9022a.o osd.o zui.o menu.o input.o
 	$(CC) $(LDFLAGS) $^ $(LDLIBS) -o $@
 
 setup.o: setup.c
@@ -41,8 +41,10 @@ menu.o: menu.c zui.h
 
 osd.o: osd.c osd.h
 
-zui.o: zui.c zui.h
+zui.o: zui.c zui.h input.h
 
 floppy.o: floppy.c floppy_img.h
 
 floppy_img.o: floppy_img.c floppy_img.h
+
+ikbd.o: ikbd.c input.h
