+++ linux/input.h
@@ -1,0 +1,26 @@
+/*
+ * input.h - input event management
+ *
+ * Copyright (c) 2022 Francois Galea <fgalea at free.fr>
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+#ifndef __INPUT_H__
+
+void input_init(void);
+
+int input_event(int timeout, int *type, int *code, int *value);
+
+#endif
+
